+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - **Cutadapt**
   - Added support for linked adapters [[#1329](https://github.com/ewels/MultiQC/issues/1329)]
   - Parse whether trimming was 5' or 3' for _Lengths of Trimmed Sequences_ plot where possible
+  - Plot filtered reads even when no filtering category is found ([#1328](https://github.com/ewels/MultiQC/issues/1328))
 - **Dragen**
   - Handled MultiQC crashing when run on single-end output from Dragen ([#1374](https://github.com/ewels/MultiQC/issues/1374))
 - **fastp**